Change of BAM slicing approach
Due to samtools merge not taking more than one region, I had to think of
a different approach. The one I finally implemented extracts the regions
from the individual subread BAM files and then those are merged. In
addition, the header of the merged file is cleaned up to only contain
the sequences that are represented in the alignments since arrow tries
to query the reference FASTA file for those. Since also the FASTA file
has been split, this will fail.